move ActionSwap to other file
