+++ ref-exchange/src/action.rs
@@ -1,0 +1,20 @@
+use near_sdk::json_types::{ValidAccountId, U128};
+use near_sdk::serde::{Deserialize, Serialize};
+
+/// Single swap action.
+#[derive(Serialize, Deserialize)]
+#[serde(crate = "near_sdk::serde")]
+pub struct SwapAction {
+    /// Pool which should be used for swapping.
+    pub pool_id: u64,
+    /// Token to swap from.
+    pub token_in: ValidAccountId,
+    /// Amount to exchange.
+    /// If amount_in is None, it will take amount_out from previous step.
+    /// Will fail if amount_in is None on the first step.
+    pub amount_in: Option<U128>,
+    /// Token to swap into.
+    pub token_out: ValidAccountId,
+    /// Required minimum amount of token_out.
+    pub min_amount_out: U128,
+}
