+++ ref-exchange/src/lib.rs
@@ -6,17 +6,18 @@ use near_contract_standards::storage_management::{
 use near_sdk::borsh::{self, BorshDeserialize, BorshSerialize};
 use near_sdk::collections::{LookupMap, UnorderedSet, Vector};
 use near_sdk::json_types::{ValidAccountId, U128};
-use near_sdk::serde::{Deserialize, Serialize};
 use near_sdk::{assert_one_yocto, env, log, near_bindgen, AccountId, PanicOnDefault, Promise};
 
 use crate::account_deposit::AccountDeposit;
+use crate::action::*;
+use crate::errors::*;
 use crate::pool::Pool;
 use crate::simple_pool::SimplePool;
 use crate::utils::check_token_duplicates;
 pub use crate::views::PoolInfo;
-use errors::*;
 
 mod account_deposit;
+mod action;
 mod errors;
 mod legacy;
 mod owner;
@@ -28,24 +29,6 @@ mod utils;
 mod views;
 
 near_sdk::setup_alloc!();
-
-/// Single swap action.
-#[derive(Serialize, Deserialize)]
-#[serde(crate = "near_sdk::serde")]
-pub struct SwapAction {
-    /// Pool which should be used for swapping.
-    pub pool_id: u64,
-    /// Token to swap from.
-    pub token_in: ValidAccountId,
-    /// Amount to exchange.
-    /// If amount_in is None, it will take amount_out from previous step.
-    /// Will fail if amount_in is None on the first step.
-    pub amount_in: Option<U128>,
-    /// Token to swap into.
-    pub token_out: ValidAccountId,
-    /// Required minimum amount of token_out.
-    pub min_amount_out: U128,
-}
 
 #[near_bindgen]
 #[derive(BorshSerialize, BorshDeserialize, PanicOnDefault)]
